Fix real root cause of splash-screen freeze: widen sdata->ptrMPK to a real pointer
The tracing added in the previous two commits pinned down exactly
where execution stops: LOAD_ReadFileASyncCallback correctly finishes
invoking its stored callback (LOAD_Callback_DriverModels via the CD
worker thread's completion path added no new corruption - the earlier
nxlink-crash fix held), but no further log line - not even the next
CTR_Main heartbeat, printed every 30 iterations - appears afterward.
Execution resumes in LOAD_TenStages case 5, immediately after case 4's
LOAD_DriverMPK call, and never returns.

Root cause: sdata->ptrMPK (include/regionsEXE.h) was declared as
`int ptrMPK` - a retail 32-bit RAM address slot, same category as the
sibling ptrLevelFile/PatchMem_Ptr fields a few lines above/below it in
the same struct. Those two were already correctly widened to real
pointer types (`struct Level *` / `void *`) in an earlier session,
but ptrMPK was missed in that audit.

LOAD_Callback_DriverModels (game/LOAD/LOAD_Callbacks.c) stores a real
host pointer into it (lqs->ptrDestination, a MEMPACK-backed buffer -
confirmed >32-bit addresses in every hardware log, e.g.
0xdebcebb30) via `(int)(uintptr_t)...`, silently truncating the upper
32 bits on Switch/AArch64 (LP64). LOAD_TenStages case 5
(game/LOAD/LOAD_TenStages.c) then reconstructs a pointer from that
truncated 32-bit value via `(uintptr_t)sdata->ptrMPK` (zero-extension
of already-lost data, not a real address) and dereferences it twice:
building sdata->PLYROBJECTLIST from it, and reading
`*(int *)(uintptr_t)sdata->ptrMPK` into gGT->mpkIcons - a read from a
wild address, which is the actual freeze/crash matching every
hardware log showing the boot stuck right here.

Fix: widen `sdata->ptrMPK` to `void *ptrMPK` (matches the sibling
fields' existing treatment; zero-cost/behavior-identical on 32-bit
PC/Vita builds since pointers are already 4 bytes there). Updates all
call sites to use it as a real pointer directly instead of round-
tripping through int/uintptr_t casts:
- LOAD_Callback_DriverModels: direct pointer assignment, no cast.
- LOAD_TenStages case 5: pointer arithmetic and NULL checks without
  uintptr_t round-trips; the *(int*) dereference is now a real pointer
  dereference into a real buffer. gGT->mpkIcons itself stays `u32`
  (see include/namespace_Main.h) since it's an offset value taken from
  inside the loaded file's own data, not a copy of this struct's
  pointer field, so it still round-trips through uintptr_t where it's
  used as an address.
- LOAD_DriverMPK (game/LOAD/LOAD_Assets.c, include/functions.h): return
  type widened from `int` to `void *` to match, since it directly
  returns sdata->ptrMPK on two paths. The return value is unused by its
  only caller (LOAD_TenStages case 4), so this is a pure signature fix,
  not a behavior change.

Verified with the local x86_64 -m64 -D__SWITCH__ syntax-check harness:
0 warnings (including -Wall -Wextra) referencing ptrMPK/LOAD_DriverMPK,
0 errors.

Diagnostic tracing from the last several commits is intentionally left
in place for this test - it should be stripped only after hardware
confirms this fixes the freeze.

+++ game/LOAD/LOAD_Callbacks.c
@@ -84,10 +84,10 @@ void LOAD_Callback_PatchMem(struct LoadQueueSlot *lqs)
 void LOAD_Callback_DriverModels(struct LoadQueueSlot *lqs)
 {
 	sdata->load_inProgress = 0;
-	// NOTE: ptrMPK is a checkpoint-tracked fixed 4-byte "retail 32-bit RAM
-	// address slot" (see platform/native_checkpoint.c); keep its type but
-	// round-trip through uintptr_t instead of casting the pointer directly.
-	sdata->ptrMPK = (int)(uintptr_t)lqs->ptrDestination;
+	// NOTE(aalhendi): ptrMPK is now a real pointer type (see regionsEXE.h) -
+	// same treatment as the sibling ptrLevelFile/PatchMem_Ptr fields, no cast
+	// truncation needed.
+	sdata->ptrMPK = lqs->ptrDestination;
 }
 
 // NOTE(aalhendi): ASM-verified NTSC-U 926 0x80031b14-0x80031b50.
